requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==6.1.3
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
django-widget-tweaks==1.4.8
pytest==7.3.1
pytest-django==4.5.2
coverage==7.2.7
